MG5integration/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Werkzeug>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pytest>=7.0.0
